==> 07_EV_GLMMS.Rmd <==
---
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, options(readr.show_col_types = FALSE))
```
 
# Statistical comparison of *Enterovirus* abundance between cases and controls in VIGR and ENDIA onset

```{r}
library(tidyverse)
library(glmmTMB)
library(gt)
```

```{r}
endia_virscan_onset <- read_rds("cache/endia_virscan_metadata_w_extra_ctrls.rds") %>% 
  filter(onset_visit == 1) %>% 
  group_by(sample_id) %>% 
  mutate(rpk = abundance / sum(abundance) * 100000) %>% 
  mutate(log_lib = log(sum(abundance) + 1e-8)) %>% 
  ungroup()

endia_for_stats <- endia_virscan_onset %>% 
  group_by(sample_id, taxon_genus) %>% 
  mutate(total_rpk = sum(rpk),
         total_abundance = sum(abundance)) %>% 
  ungroup() %>% 
  distinct(taxon_genus, sample_id, .keep_all = TRUE) %>% 
  mutate(across(c(infant_HLA, infant_id, mother_id, maternal_T1D, deidentified_nest_id_new, infant_sex), as.factor),
        age_sample_collection_month = age_at_sample_collection_days / 365 * 12,
        age_c = scale(age_at_sample_collection_days),
        log_total_rpk = log(total_rpk + 1),
        Condition = factor(condition, levels = c("Control", "Case"), labels = c("Control", "Case")),
        infant_HLA = fct_relevel(infant_HLA, "DRXX"))
```

```{r}
endia_ev_for_stats <- endia_for_stats %>% filter(taxon_genus == "Enterovirus")

endia_ev_model_poisson <- glmmTMB(formula = case ~ log_total_rpk + infant_HLA + age_sample_collection_month + infant_sex + maternal_T1D + (1|deidentified_nest_id_new) + (1|mother_id),
       family = poisson(link = "log"),
        weights = weightTEDDY,
       data = endia_ev_for_stats)

endia_ev_model_poisson_tidy <- endia_ev_model_poisson %>%
    broom.mixed::tidy(effect = "fixed", conf.int = TRUE, exponentiate = TRUE) %>% 
    add_column(cohort = "ENDIA")
```

```{r}
vigr_virscan_metadata <- read_rds("cache/vigr_virscan_metadata.rds") %>% 
 group_by(sample_id) %>% 
  mutate(rpk = abundance / sum(abundance) * 100000,
         log_lib = log(sum(abundance) + 1e-8)) %>% 
  ungroup() %>% 
  mutate(Nest = str_extract(sample_id, "\\d+"))

vigr_for_stats <- vigr_virscan_metadata %>% 
  group_by(sample_id, taxon_genus) %>%
  mutate(total_rpk = sum(rpk),
         total_abundance = sum(abundance)) %>% 
  ungroup() %>% 
  distinct(taxon_genus, sample_id, .keep_all = TRUE) %>% 
  mutate(case = ifelse(Condition == "Case", 1, 0),
        maternal_T1D = ifelse(maternal_T1D == "T1D", 1, 0),
        log_total_rpk = log(total_rpk + 1),
        across(c(HLA_Status, Sex, Nest, maternal_T1D, My_name), as.factor)) %>% 
  mutate(Condition = factor(Condition, levels = c("Control", "Case"), labels = c("Control", "Case")))

vigr_ev_for_stats <- vigr_for_stats %>% filter(taxon_genus == "Enterovirus")

vigr_ev_model_poisson <- glmmTMB(case ~ log_total_rpk + HLA_Status + Age + Sex + maternal_T1D + (1|Nest),
                                family = poisson(link = "log"),
                                data = vigr_ev_for_stats)

vigr_ev_model_poisson_tidy <- vigr_ev_model_poisson %>%  
  broom.mixed::tidy(effect = "fixed", conf.int = TRUE, exponentiate = TRUE) %>% 
    add_column(cohort = "VIGR")
```

```{r}
tidy_models_data <- rbind(endia_ev_model_poisson_tidy, vigr_ev_model_poisson_tidy) %>% 
    dplyr::rename(RR = estimate,
               `CI lower` = conf.low,
               `CI upper` = conf.high) %>% 
    select(term, RR, `CI lower`, `CI upper`, p.value, cohort) %>% 
    filter(term != "(Intercept)")
```

GLM models output for ENDIA and VIGR table

```{r, echo = FALSE}
tidy_models_data_table <- tidy_models_data %>%
  mutate(term = case_match(term, !!!c("log_total_rpk" ~ "log(EV)"), .default = term)) %>% 
  select(-cohort) %>% 
  gt(rowname_col = "term") %>%
  tab_header(
    title = "Summary of GLM model tests"
  ) %>%
  cols_label(
    term = "Predictor",
    RR = "RR",
    `CI lower` = "CI Lower",
    `CI upper` = "CI Upper",
    p.value = "Pr(>|z|)"
  ) %>%
  fmt_number(
    columns = c(RR, `CI lower`, `CI upper`, p.value),
    decimals = 3
  ) %>%
  cols_align(
    align = "left",
    columns = everything()
  ) %>% 
  tab_row_group(
    label = "ENDIA",
    rows = 1:7
  ) %>% 
  tab_row_group(
    label = "VIGR",
    rows = 8:13
  ) %>% 
  tab_style(
    style = cell_fill(color = "lightgrey"),
    locations = cells_row_groups(groups = c("ENDIA", "VIGR"))
  )

as_raw_html(tidy_models_data_table)

```

```{r, echo = FALSE}
gtsave(tidy_models_data_table, "figures/glm_results_vigr_endia_onset.png")
```
